refactor (enhance core) : add enhancore with power init config of create Element
