+++ src/lib/core/index.ts
@@ -1,0 +1,2 @@
+export { EnhanceCore } from './enhance'
+export { Core } from './base'
